Rename ReceiveInfo to Sep31Info
