+++ api-schema/src/main/java/org/stellar/anchor/api/exception/Sep31MissingFieldException.java
@@ -1,17 +1,17 @@
 package org.stellar.anchor.api.exception;
 
-import org.stellar.anchor.api.sep.operation.ReceiveInfo;
+import org.stellar.anchor.api.sep.operation.Sep31Info;
 
 /** Thrown when a SEP-31 transaction is missing required fields. */
 public class Sep31MissingFieldException extends AnchorException {
-  private final ReceiveInfo.Fields missingFields;
+  private final Sep31Info.Fields missingFields;
 
-  public Sep31MissingFieldException(ReceiveInfo.Fields missingFields) {
+  public Sep31MissingFieldException(Sep31Info.Fields missingFields) {
     super();
     this.missingFields = missingFields;
   }
 
-  public ReceiveInfo.Fields getMissingFields() {
+  public Sep31Info.Fields getMissingFields() {
     return missingFields;
   }
 }
