+++ api-schema/src/main/java/org/stellar/anchor/api/sep/operation/Sep31Info.java
@@ -6,7 +6,7 @@ import lombok.Data;
 import org.stellar.anchor.api.asset.AssetInfo;
 
 @Data
-public class ReceiveInfo {
+public class Sep31Info {
   Boolean enabled = false;
 
   ReceiveOperation receive;
