+++ api-schema/src/main/java/org/stellar/anchor/api/sep/sep31/Sep31InfoResponse.java
@@ -3,8 +3,8 @@ package org.stellar.anchor.api.sep.sep31;
 import com.google.gson.annotations.SerializedName;
 import java.util.Map;
 import lombok.Data;
-import org.stellar.anchor.api.sep.operation.ReceiveInfo;
 import org.stellar.anchor.api.sep.operation.Sep12Operation;
+import org.stellar.anchor.api.sep.operation.Sep31Info;
 
 /**
  * The response body of the /info endpoint of the SEP-31.
@@ -40,6 +40,6 @@ public class Sep31InfoResponse {
     Long maxAmount;
 
     Sep12Operation sep12;
-    ReceiveInfo.Fields fields;
+    Sep31Info.Fields fields;
   }
 }
