+++ core/src/test/java/org/stellar/anchor/sep31/PojoSep31Transaction.java
@@ -4,7 +4,7 @@ import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 import lombok.Data;
-import org.stellar.anchor.api.sep.operation.ReceiveInfo;
+import org.stellar.anchor.api.sep.operation.Sep31Info;
 import org.stellar.anchor.api.shared.FeeDescription;
 import org.stellar.anchor.api.shared.FeeDetails;
 import org.stellar.anchor.api.shared.StellarId;
@@ -35,7 +35,7 @@ public class PojoSep31Transaction implements Sep31Transaction {
   String quoteId;
   String clientDomain;
   String clientName;
-  ReceiveInfo.Fields requiredInfoUpdates;
+  Sep31Info.Fields requiredInfoUpdates;
   Map<String, String> fields;
   Boolean refunded;
   PojoSep31Refunds refunds;
