+++ platform/src/main/java/org/stellar/anchor/platform/data/JdbcSep31Transaction.java
@@ -11,7 +11,7 @@ import org.hibernate.annotations.JdbcTypeCode;
 import org.hibernate.type.SqlTypes;
 import org.springframework.beans.BeanUtils;
 import org.stellar.anchor.SepTransaction;
-import org.stellar.anchor.api.sep.operation.ReceiveInfo;
+import org.stellar.anchor.api.sep.operation.Sep31Info;
 import org.stellar.anchor.api.shared.StellarId;
 import org.stellar.anchor.sep31.Sep31Refunds;
 import org.stellar.anchor.sep31.Sep31Transaction;
@@ -96,7 +96,7 @@ public class JdbcSep31Transaction extends JdbcSepTransaction
   // Ignored by JPA and Gson
   @SerializedName("required_info_updates")
   @Transient
-  ReceiveInfo.Fields requiredInfoUpdates;
+  Sep31Info.Fields requiredInfoUpdates;
 
   @Access(AccessType.PROPERTY)
   @Column(name = "requiredInfoUpdates")
@@ -106,7 +106,7 @@ public class JdbcSep31Transaction extends JdbcSepTransaction
 
   public void setRequiredInfoUpdatesJson(String requiredInfoUpdatesJson) {
     if (requiredInfoUpdatesJson != null) {
-      this.requiredInfoUpdates = gson.fromJson(requiredInfoUpdatesJson, ReceiveInfo.Fields.class);
+      this.requiredInfoUpdates = gson.fromJson(requiredInfoUpdatesJson, Sep31Info.Fields.class);
     }
   }
